Enable comment liquid tag and alias devcomment to it
