+++ app/assets/stylesheets/ltags/LiquidTags.scss
@@ -2,7 +2,7 @@
 @import 'TweetTag';
 @import 'YoutubeTag';
 @import 'GithubTag';
-@import 'DevCommentTag';
+@import 'CommentTag';
 @import 'LinkTag';
 @import 'PodcastTag';
 @import 'UserTag';
