+++ app/liquid_tags/null_tag.rb
@@ -4,7 +4,7 @@ class NullTag < Liquid::Block
   end
 end
 
-disabled_tags = %w[assign break capture case comment cycle decrement for if ifchanged include increment unless tablerow]
+disabled_tags = %w[assign break capture case cycle decrement for if ifchanged include increment unless tablerow]
 
 disabled_tags.each do |tag|
   Liquid::Template.register_tag(tag, NullTag)
